refactor(core): split pipeline-style-families tests by concern
Break the 1604-line monolith into a directory of focused suites
(geom plumbing, resolveStyleScale units, grouping, scale policy,
guides/legends) so review and retrieval match one responsibility
per file. Assertions and it-titles are preserved; fixtures share
only the viewport.
