+++ packages/core/tests/pipeline-style-families/fixtures.ts
@@ -1,0 +1,2 @@
+/** Shared fixtures for mapped style-family pipeline tests. */
+export const viewport = { width: 640, height: 400 };
